+++ R/ris_search_dok.R
@@ -59,7 +59,6 @@ ris_search_dok <- function(
   search_decision_text = TRUE,
   search_legal_principles = TRUE,
   echo = FALSE,
-  max_pages = Inf,
   base_url = ris_base_url()
 ) {
   checkmate::assert_string(
@@ -81,7 +80,6 @@ ris_search_dok <- function(
     search_decision_text = search_decision_text,
     search_legal_principles = search_legal_principles,
     echo = echo,
-    max_pages = max_pages,
     base_url = base_url
   )
 }