+++ R/ris_search_lvwg.R
@@ -74,7 +74,6 @@ ris_search_lvwg <- function(
   search_decision_text = TRUE,
   search_legal_principles = TRUE,
   echo = FALSE,
-  max_pages = Inf,
   base_url = ris_base_url()
 ) {
   federal_state <- ris_normalize_federal_state(federal_state)
@@ -93,7 +92,6 @@ ris_search_lvwg <- function(
     search_decision_text = search_decision_text,
     search_legal_principles = search_legal_principles,
     echo = echo,
-    max_pages = max_pages,
     base_url = base_url
   )
 }