+++ R/ris_search_pvak.R
@@ -57,7 +57,6 @@ ris_search_pvak <- function(
   search_decision_text = TRUE,
   search_legal_principles = TRUE,
   echo = FALSE,
-  max_pages = Inf,
   base_url = ris_base_url()
 ) {
   checkmate::assert_string(
@@ -79,7 +78,6 @@ ris_search_pvak <- function(
     search_decision_text = search_decision_text,
     search_legal_principles = search_legal_principles,
     echo = echo,
-    max_pages = max_pages,
     base_url = base_url
   )
 }