+++ R/ris_search_vwgh.R
@@ -98,7 +98,6 @@ ris_search_vwgh <- function(
   search_decision_text = TRUE,
   search_legal_principles = TRUE,
   echo = FALSE,
-  max_pages = Inf,
   base_url = ris_base_url()
 ) {
   # Delegate to the generic case law search with application fixed to "Vwgh".
@@ -116,7 +115,6 @@ ris_search_vwgh <- function(
     search_decision_text = search_decision_text,
     search_legal_principles = search_legal_principles,
     echo = echo,
-    max_pages = max_pages,
     base_url = base_url
   )
 }